requirements.txt
Babel==2.8.0
beautifulsoup4==4.6.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.8
django-appconf==1.0.3
django-countries==5.5
django-crispy-forms==1.8.1
django-eventtools==0.9.11
django-extensions==2.2.5
django-filter==2.2.0
django-froala-editor==3.1.0
django-haystack==2.8.1
django-imagekit==4.0.2
django-intl-tel-input==0.3.1
django-js-asset==1.2.2
django-machina==1.0.2
django-modelcluster==5.0
django-money==1.0
django-mptt==0.10.0
django-phonenumber-field==4.0.0
django-settings-export==1.2.1
django-storages==1.9
django-taggit==1.2.0
django-treebeard==4.3
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
draftjs-exporter==2.1.7
ecdsa==0.15
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
html5lib==1.0.1
idna==2.8
isoweek==1.3.3
lxml==4.5.0
markdown2==2.3.8
mailer==0.8.1
oauthlib==3.1.0
openpyxl==3.0.3
phonenumberslite==8.11.3
pilkit==2.0
Pillow==6.2.1
protobuf==3.11.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodomex==3.9.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
shortuuid==0.5.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stream-django==1.6.1
stream-python==3.1.1
Unidecode==1.1.1
urllib3==1.25.7
wagtail==2.7
wagtail-blog==2.1.0
wagtailgeowidget==5.0.2
wagtailleafletwidget==1.0.1
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.0.1
Willow==1.3
